requirements.txt
selenium
webdriver_manager
PyGithub